+++ data_gen_templates/rules.py
@@ -90,9 +90,9 @@ class Rules:
                 else:
                     mons.add(f"mon_{spec.pre_evolution.species}")
                 spec = new_spec
-
+        bag = items.items["bag"].label
         def hm_rule(state: CollectionState) -> bool:
-            if not (state.has(hm, player) and self.common_rules[f"{hm.name.lower()}_badge"](state)):
+            if not (state.has_all([hm, bag], player) and self.common_rules[f"{hm.name.lower()}_badge"](state)):
                 return False
             if state.has_any(mons, player):
                 return True